pyproject.toml
[project]
name = "Data_Tool_Pack_Py"
version = "0.1.0"
description = "Data_Tool_Pack_Py"
requires-python = ">=3.10,<3.11"
dependencies = [
    "black>=24.10.0",
    "databricks-cli>=0.18.0",
    "databricks-sql-connector>=3.7.1",
    "datetime>=5.5",
    "fitz>=0.0.1.dev2",
    "flake8>=7.1.1",
    "geopy>=2.4.1",
    "imageio[ffmpeg]>=2.36.1",
    "impyla>=0.19.0",
    "ipykernel>=6.29.5",
    "isort>=5.13.2",
    "jaydebeapi>=1.2.3",
    "jira>=3.8.0",
    "looker-sdk>=24.20.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "pydrive>=1.3.1",
    "pygsheets>=2.0.6",
    "pytesseract>=0.3.13",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "snowflake-connector-python>=3.12.4",
    "streamlit>=1.41.1",
    "tabula>=1.0.5",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    "types-requests>=2.32.0.20241016",
]

[tool.flake8]
max-line-length = 120
max-complexity = 15
extend-ignore = ["E203", "W503"]

[tool.isort]
profile = "black"
lines_between_types = 0
lines_between_sections = 1
sort_relative_in_force_sorted_sections = false
known_third_party = ["brand_tools", "budgets", "breakout_gcr_pl_mapping", "breakout_schemas", "config", "config_utils", "config_tests", "config_data_quality_validation", "concur_schemas", "ap_vendor_split_hc_us"]
sections= ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["dotenv", "test_tools"]
no_lines_before = ["FIRSTPARTY"]
force_sort_within_sections = false

[tool.mypy]
ignore_missing_imports = true
show_error_codes = true
